FIX: Redraw graph when switching to and from visualisation states
If you switch between unit and tutorial, you'll notice that it wouldn't
redraw the graph the second time you switch to tutorial. This is because
the selectedTutorial is not changed when you go back. Now, we enforce
the selectedTutorial/selectedTask to be updated each time the button group
state (selectedType) is changed.