FIX: Rework submission wizard for needs help
Students were getting confused as to how to upload work
after it was submitted as 'need help'. Previously this
required student to click the "ready to mark", then
click "ready to mark" again in the file uploader ( as
it defaulted to the previous state). Now this updates
the state of the task when the status is changed in the
task viewer.