FIX: Force max quality star ratings to update on task change
As per https://angular-ui.github.io/bootstrap/#/rating the max
value is not watched so we can't just provide a dynamic max star
rating value here.

Instead we can use an ng-if to force the element to be destroyed
when the max_quality_pts changes value, and then a $timeout to
force it to be created again.