FIX: Ensure task with correct definitions are loaded
For the task def list, if I start with

    student_20/A1

and switch to

    student_7/A8

nothing will happen as tasks of definition "A8" have not
been loaded so it thinks that none of these tasks exist.

We need to ensure that these are actually loaded before it
searches for these tasks.

Removing the filter on unit.fillWithUnStartedTasks allows
this to not fail. Filtering out by target grade on a task
where the student's target grade does not match that of
the student's target grade fails the search as, once again
that student's task cannot be found as it gets filtered
out. Not sure why this was there to begin with, will
need to talk with @macite to see if this causes any potential
issues.