FIX: Fix duplicate options in ng-options
Under angular.js:26629 in the updateOptions function, the
currentElement variable will be assigned to the generated
comment of the ng-if (i.e., <!-- ngIf: !isTaskDefMode -->)
and not the select <option> element.

This is a bug with angular by having a first, default option
in a select with ng-options -- it *must not* have an ng-if
clause on it or duplicate entries are added.