Update task-comment-composer.component.ts
Removed redundant code which causes errors.

NgOnChanges now only looks to see if the task is changed. If the task is changed, it loads its content based on its key. Every keystroke is saved in the local storage; therefore, we draft is loaded, we can pull this from the storage straight away.

When its empty, it already deletes the whole key, so the content is no longer loaded.